Move DomainEntity processing to different service.
