+++ src/main/java/no/ndla/taxonomy/service/task/Fetcher.java
@@ -7,10 +7,8 @@
 
 package no.ndla.taxonomy.service.task;
 
-import no.ndla.taxonomy.domain.CustomField;
-import no.ndla.taxonomy.domain.DomainEntity;
-import no.ndla.taxonomy.domain.EntityWithMetadata;
-import no.ndla.taxonomy.domain.Metadata;
+import no.ndla.taxonomy.domain.*;
+import no.ndla.taxonomy.service.ChangelogService;
 import no.ndla.taxonomy.service.CustomFieldService;
 import no.ndla.taxonomy.service.DomainEntityHelperService;
 import no.ndla.taxonomy.service.exceptions.EntityNotFoundException;
@@ -23,7 +21,6 @@ import java.util.stream.Collectors;
 
 public class Fetcher extends Task<DomainEntity> {
     private DomainEntityHelperService domainEntityHelperService;
-    private CustomFieldService customFieldService;
     private URI publicId;
     private boolean addIsPublishing;
 
@@ -32,10 +29,6 @@ public class Fetcher extends Task<DomainEntity> {
 
     public void setDomainEntityHelperService(DomainEntityHelperService domainEntityHelperService) {
         this.domainEntityHelperService = domainEntityHelperService;
-    }
-
-    public void setCustomFieldService(CustomFieldService customFieldService) {
-        this.customFieldService = customFieldService;
     }
 
     public void setPublicId(URI publicId) {
@@ -47,39 +40,8 @@ public class Fetcher extends Task<DomainEntity> {
     }
 
     @Override
-    @Transactional
     protected Optional<DomainEntity> execute() {
-        DomainEntity entity = domainEntityHelperService.getEntityByPublicId(this.publicId);
-        if (entity instanceof EntityWithMetadata) {
-            EntityWithMetadata entityWithMetadata = (EntityWithMetadata) entity;
-            if (addIsPublishing && !cleanUp) {
-                if (!entityWithMetadata.getMetadata().getCustomFieldValues().stream()
-                        .map(customFieldValue -> customFieldValue.getCustomField().getKey())
-                        .collect(Collectors.toList()).contains(CustomField.IS_PUBLISHING)) {
-                    customFieldService.setCustomField(entityWithMetadata.getMetadata(), CustomField.IS_PUBLISHING,
-                            "true");
-                }
-                return Optional.of(entity);
-            }
-            if (cleanUp) {
-                Metadata metadata = entityWithMetadata.getMetadata();
-                unsetCustomField(metadata, CustomField.IS_PUBLISHING);
-                unsetCustomField(metadata, CustomField.REQUEST_PUBLISH);
-                unsetCustomField(metadata, CustomField.IS_CHANGED);
-            }
-        }
-        return Optional.ofNullable(entity);
-    }
-
-    private void unsetCustomField(Metadata metadata, String customfield) {
-        metadata.getCustomFieldValues().forEach(customFieldValue -> {
-            if (customFieldValue.getCustomField().getKey().equals(customfield)) {
-                try {
-                    customFieldService.unsetCustomField(customFieldValue.getId());
-                } catch (EntityNotFoundException e) {
-                    // Already deleted. Do nothing.
-                }
-            }
-        });
+        return domainEntityHelperService.getProcessedEntityByPublicId(this.publicId, this.addIsPublishing,
+                this.cleanUp);
     }
 }
